+++ packages/mochi-rest/src/schemas/index.ts
@@ -10,3 +10,4 @@ export * from "./watchlist";
 export * from "./pagination";
 export * from "./moniker";
 export * from "./pay-link";
+export * from "./stats";